Rework parsing (+backwards compatibility).
After I've decided to completely ignore the iterations thing
(reruning all steps multiple times) I decided to add a backwards
compatibility. I've added a new parameter in the `KodgenSettings.toml` -
`shouldFailCodeGenerationOnClangErrors` which is `false` by default.
When this setting is turned off we use the old parsing algorithm
(existed before this branch, ignores parsing errors), when it's
turned on we use the new approach where we fail on any Clang error
that occur during the parsing.