Retry parsing failed files after all files.
While running parsing for the tests I've noticed that there is a
situation when 1 GENERATED macro is not captured in the pre-parsing
step which fails the parsing step. I also noticed that this error
might not occur due to the fact that pre-parsing step writes
macros in parallel. Thus, I've made it so that the pre-parsing step
writes found GENERATED macros in a single-threaded manner after the
pre-parsing step but before the parsing step. Although this is indeed an
improvement, it does not solved the issue where 1 GENERATED macro was
not found. I then remembered about "iterations" thing and added a cycle.
So now, files that failed the parsing step (only failed files) will be
queued to be parsed again on the next cycle iteration. With this change,
when we rerun the pre-parsing step for failed files we have enough info
to find all GENERATED macros. Of cource I added a check where if the
number of failed files stays the same we break out of the cycle and
show an error.